+++ test/test-runner.sh
@@ -12,4 +12,6 @@ LOG_LEVEL=error ./node_modules/.bin/mocha ./test/LOG_LEVEL_error.test.js
 
 SECURITY_TEST=true ./node_modules/.bin/mocha ./test/SECURITY_TEST_true.test.js
 
+RELEASE_ENV=uat ./node_modules/.bin/mocha ./test/RELEASE_ENV_uat.test.js
 RELEASE_ENV=qa ./node_modules/.bin/mocha ./test/RELEASE_ENV_qa.test.js
+RELEASE_ENV=production ./node_modules/.bin/mocha ./test/RELEASE_ENV_production.test.js